src/test/java/com/example/alg/arraypkg/DisappearedNumbers.java
package com.example.alg.arraypkg;

import java.util.ArrayList;
import java.util.Collection;
import java.util.List;
import java.util.Map;
import java.util.function.Function;
import java.util.stream.Collectors;
import java.util.stream.IntStream;

public class DisappearedNumbers {
    /**
     * Given an array of integers where 1 ≤ a[i] ≤ n (n = size of array), some elements appear twice and others appear once.

     Find all the elements of [1, n] inclusive that do not appear in this array.

     Could you do it without extra space and in O(n) runtime? You may assume the returned list does not count as extra space.

     Example:

     Input:
     [4,3,2,7,8,2,3,1]

     Output:
     [5,6]
     */
    public List<Integer> findDisappearedNumbers(int[] nums) {

        final Map<Integer, Integer> map = IntStream.rangeClosed(1, nums.length).boxed()
                .collect(Collectors.toMap(Function.identity(), Function.identity()));

        for (int num : nums) {
            map.remove(num);
        }
        return new ArrayList<>(map.values());
    }

    public List<Integer> findDisappearedNumbers1(int[] nums) {
        List<Integer> ret = new ArrayList<Integer>();

        for(int i = 0; i < nums.length; i++) {
            int val = Math.abs(nums[i]) - 1;
            if(nums[val] > 0) {
                nums[val] = -nums[val];
            }
        }

        for(int i = 0; i < nums.length; i++) {
            if(nums[i] > 0) {
                ret.add(i+1);
            }
        }
        return ret;
    }
}
